requirements.txt
Django==2.2.19
django-leaflet==0.28.3
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
